backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pgvector==0.3.5
celery[redis]==5.4.0
redis==5.1.1
python-multipart==0.0.12
pillow==10.4.0
pdf2image==1.17.0
pytesseract==0.3.13
httpx==0.27.2
sentence-transformers==3.1.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
watchdog==5.0.3
python-magic==0.4.27
aiofiles==24.1.0
psycopg2-binary==2.9.9
